Display config errors and add number field.
