+++ src/panels/ConfigPanel/NumberConfig.tsx
@@ -1,30 +1,29 @@
-import { useRef } from "react";
 import ConfigField, { ConfigProps } from "./ConfigField";
 
-export default function TextAreaConfig({
+const NumberConfig = ({
   config,
   selectedConfig,
   setSelectedConfig,
   setNewValue,
-}: ConfigProps): JSX.Element {
-  const inputRef = useRef<HTMLTextAreaElement>(null);
-
+}: ConfigProps): JSX.Element => {
   return (
-    <ConfigField<string>
+    <ConfigField<number>
       config={config}
       selectedConfig={selectedConfig}
       setSelectedConfig={setSelectedConfig}
       setNewValue={setNewValue}
       input={(value) => (
-        <textarea
-          ref={inputRef}
+        <input
+          type="number"
           value={value}
           onFocus={() => setSelectedConfig(config)}
           onChange={(e) => {
             setNewValue(config.name, e.target.value);
           }}
-        ></textarea>
+        />
       )}
     />
   );
-}
+};
+
+export default NumberConfig;
